+++ k8s/kubernetes_workload.go
@@ -95,13 +95,13 @@ func checkIfPodHasStatusRunningInNamespace(podNameStartsWith string, namespace s
 				os.Exit(1)
 
 			case v1.PodPending:
-				makeup.Print("The Pod " + pod.Name + "h in pending but should be running.")
+				makeup.Print("The Pod " + pod.Name + " in pending but should be running.")
 				return false
 			case v1.PodSucceeded:
-				makeup.Print("The Pod " + pod.Name + "h has succeeded but should be running.")
+				makeup.Print("The Pod " + pod.Name + " has succeeded but should be running.")
 				return false
 			case v1.PodUnknown:
-				makeup.Print("The Pod " + pod.Name + "h has an unknown status but should be running.")
+				makeup.Print("The Pod " + pod.Name + " has an unknown status but should be running.")
 				return false
 			default:
 				return false
@@ -190,6 +190,13 @@ func GetDynamicKubernetesClient() *dynamic.DynamicClient {
 }
 
 /*
+Wait for a Kubernetes resource to reach either a desired or failed state.
+
+Namespace
+Name: name of the object to wait for, e.g. name of the backup
+The desiredConditionsMap contains the conditions to indicate success while
+the failedConditionsMap contains the conditions to indicate failure.
+
 Example:
 
 gvr := schema.GroupVersionResource{Group: "backups.anynines.com", Version: "v1beta3", Resource: "backups"}
@@ -197,13 +204,25 @@ gvr := schema.GroupVersionResource{Group: "backups.anynines.com", Version: "v1be
 desiredConditionsMap := make(map[string]interface{})
 desiredConditionsMap["reason"] = "Complete"
 desiredConditionsMap["status"] = "True"
+
+failedConditionsMap := make(map[string]interface{})
+failedConditionsMap["reason"] = "PermanentlyFailed"
+failedConditionsMap["status"] = "True"
+
+TODO
+  - Refactor using WaitForKubernetesResourceWithFunction
+  - Rename WaitForKubernetesResourceWithFunction to WaitForKubernetesResource
 */
-func WaitForKubernetesResource(namespace string, gvr schema.GroupVersionResource, desiredConditionsMap map[string]interface{}) error {
+func WaitForKubernetesResource(namespace, name string, gvr schema.GroupVersionResource, desiredConditionsMap map[string]interface{}, failedConditionsMap map[string]interface{}) error {
+
+	listOptions := metav1.ListOptions{
+		FieldSelector: fmt.Sprintf("metadata.name=%s", name),
+	}
 
 	dynamicClient := GetDynamicKubernetesClient()
 
 	// Watch for changes in Backup resources.
-	watchInterface, err := dynamicClient.Resource(gvr).Namespace(namespace).Watch(context.TODO(), metav1.ListOptions{})
+	watchInterface, err := dynamicClient.Resource(gvr).Namespace(namespace).Watch(context.TODO(), listOptions)
 	if err != nil {
 		makeup.ExitDueToFatalError(err, fmt.Sprintf("Can't create dynamic WatchInterface to watch Kubernetes resource %v.", gvr))
 	}
@@ -222,6 +241,7 @@ func WaitForKubernetesResource(namespace string, gvr schema.GroupVersionResource
 			}
 
 			if makeup.Verbose {
+				fmt.Print("Event object:")
 				pretty.Print(backup)
 			}
 
@@ -263,6 +283,7 @@ func WaitForKubernetesResource(namespace string, gvr schema.GroupVersionResource
 			}
 
 			conditionsAreMet := false
+			failedConditionsAreMet := false
 
 			for _, condition := range conditions {
 				makeup.PrintVerbose(fmt.Sprintf("Investigating condition %v of conditions\n", condition))
@@ -281,9 +302,15 @@ func WaitForKubernetesResource(namespace string, gvr schema.GroupVersionResource
 					conditionsAreMet = true
 					break
 				}
+
+				if failedConditionsMap != nil && ConditionsAreMet(condMap, failedConditionsMap) {
+					failedConditionsAreMet = true
+					break
+				}
 			}
 
 			//TODO The conditionsAreMet variable is not necessary but increases readability. Does it?
+			// No it doesn't. Code here could also be put in the above if clause (if ConditionsAreMet ...)
 			if conditionsAreMet {
 				//makeup.PrintCheckmark("Operation complete for resource: " + backup.GetName())
 				return nil
@@ -294,7 +321,64 @@ func WaitForKubernetesResource(namespace string, gvr schema.GroupVersionResource
 				}
 			}
 
+			if failedConditionsAreMet {
+				errorMessage := fmt.Sprintf("waiting for Kubernetes resource %v in namespace %s has failed. Resource reached failed state", gvr, namespace)
+				if makeup.Verbose {
+					makeup.PrintWarning(errorMessage)
+				}
+				return errors.New(errorMessage)
+			}
+
 			continue
+		}
+	}
+	return errors.New("expected conditions have not been met")
+}
+
+/*
+name refers to the metadata.name value of the object of interest.
+
+waitLonger is a function describing what to wait for covering both success and failure scenarios.
+It returns true if waiting shall go on and false if the awaited event has happened.g
+*/
+func WaitForKubernetesResourceWithFunction(namespace, name string, gvr schema.GroupVersionResource, waitLonger func(object *m1u.Unstructured) bool) error {
+
+	listOptions := metav1.ListOptions{
+		FieldSelector: fmt.Sprintf("metadata.name=%s", name),
+	}
+
+	dynamicClient := GetDynamicKubernetesClient()
+
+	// Watch for changes in Backup resources.
+	watchInterface, err := dynamicClient.Resource(gvr).Namespace(namespace).Watch(context.TODO(), listOptions)
+	if err != nil {
+		makeup.ExitDueToFatalError(err, fmt.Sprintf("Can't create dynamic WatchInterface to watch Kubernetes resource %v.", gvr))
+	}
+
+	makeup.PrintVerbose(fmt.Sprintf("Watching for resources: %v", gvr))
+
+	var goOn bool
+
+	for event := range watchInterface.ResultChan() {
+		switch event.Type {
+		case watch.Error:
+			makeup.ExitDueToFatalError(err, "A watch.Error occurred watching the resource.")
+		case watch.Added, watch.Modified:
+			object, ok := event.Object.(*m1u.Unstructured)
+			if !ok {
+				makeup.ExitDueToFatalError(nil, "Could not cast to Unstructured")
+			}
+
+			if makeup.Verbose {
+				fmt.Print("Event object:")
+				pretty.Print(object)
+			}
+
+			goOn = waitLonger(object)
+
+			if !goOn {
+				return nil
+			}
 		}
 	}
 	return errors.New("expected conditions have not been met")